requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20230802.0=h313beb8_2
altair=5.0.1=py310hca03da5_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
appnope=0.1.2=py310hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h1a28f6b_0
arrow-cpp=14.0.2=hc7aafb3_1
astroid=2.14.2=py310hca03da5_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py310hca03da5_0
attrs=23.1.0=py310hca03da5_0
aws-c-auth=0.6.19=h80987f9_0
aws-c-cal=0.5.20=h80987f9_0
aws-c-common=0.8.5=h80987f9_0
aws-c-compression=0.2.16=h80987f9_0
aws-c-event-stream=0.2.15=h313beb8_0
aws-c-http=0.6.25=h80987f9_0
aws-c-io=0.13.10=h80987f9_0
aws-c-mqtt=0.7.13=h80987f9_0
aws-c-s3=0.1.51=h80987f9_0
aws-c-sdkutils=0.1.6=h80987f9_0
aws-checksums=0.1.13=h80987f9_0
aws-crt-cpp=0.18.16=h313beb8_0
aws-sdk-cpp=1.10.55=h313beb8_0
babel=2.11.0=py310hca03da5_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=py310hca03da5_0
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.3.7=py310hbda83bc_0
brotli=1.0.9=h80987f9_8
brotli-bin=1.0.9=h80987f9_8
brotli-python=1.0.9=py310h313beb8_8
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.3.11=hca03da5_0
cachetools=5.3.3=py310hca03da5_0
certifi=2024.2.2=py310hca03da5_0
cffi=1.16.0=py310h80987f9_1
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py310hca03da5_0
colorama=0.4.6=py310hca03da5_0
comm=0.2.1=py310hca03da5_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.8=py310hca03da5_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=7.1.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=0.8.3=pyhd3eb1b0_0
flaml=2.1.2=pypi_0
fonttools=4.52.4=pypi_0
freetype=2.12.1=h1192e45_0
frozendict=2.4.4=pypi_0
gflags=2.2.2=h313beb8_1
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.37=py310hca03da5_0
glog=0.5.0=h313beb8_1
grpc-cpp=1.48.2=hc60591f_4
gtest=1.14.0=h48ca7d4_1
h11=0.14.0=pypi_0
html5lib=1.1=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
icu=73.1=h313beb8_0
idna=3.7=py310hca03da5_0
ipykernel=6.28.0=py310hca03da5_0
ipython=8.20.0=py310hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.8.1=py310hca03da5_0
isort=5.9.3=pyhd3eb1b0_0
jedi=0.18.1=py310hca03da5_1
jinja2=3.1.3=py310hca03da5_0
jpeg=9e=h80987f9_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py310hca03da5_0
jsonschema-specifications=2023.7.1=py310hca03da5_0
jupyter-lsp=2.2.0=py310hca03da5_0
jupyter_client=8.6.0=py310hca03da5_0
jupyter_core=5.5.0=py310hca03da5_0
jupyter_events=0.8.0=py310hca03da5_0
jupyter_server=2.10.0=py310hca03da5_0
jupyter_server_terminals=0.4.4=py310hca03da5_1
jupyterlab=4.0.11=py310hca03da5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py310hca03da5_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=hf3e1bf2_1
lazy-object-proxy=1.10.0=py310h80987f9_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h80987f9_8
libbrotlidec=1.0.9=h80987f9_8
libbrotlienc=1.0.9=h80987f9_8
libcurl=8.7.1=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h80987f9_3
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h1a28f6b_0
libssh2=1.11.0=h3e2b118_0
libthrift=0.15.0=h73c2103_2
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lxml=5.2.2=pypi_0
lz4-c=1.9.4=h313beb8_1
markdown-it-py=2.2.0=py310hca03da5_1
markupsafe=2.1.3=py310h80987f9_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.6=py310hca03da5_0
mccabe=0.7.0=pyhd3eb1b0_0
mdurl=0.1.0=py310hca03da5_0
mistune=2.0.4=py310hca03da5_0
multitasking=0.0.11=pypi_0
nbclient=0.8.0=py310hca03da5_0
nbconvert=7.10.0=py310hca03da5_0
nbformat=5.9.2=py310hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=py310hca03da5_0
notebook=7.0.8=py310hca03da5_0
notebook-shim=0.2.3=py310hca03da5_0
numexpr=2.8.7=py310hecc3335_0
numpy=1.26.4=py310h3b2db8e_0
numpy-base=1.26.4=py310ha9811e2_0
openai=1.30.4=pypi_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.0.13=h1a28f6b_2
orc=1.7.4=hdca1487_1
overrides=7.4.0=py310hca03da5_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
peewee=3.17.5=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.3.0=py310h80987f9_0
pip=24.0=py310hca03da5_0
platformdirs=4.2.2=pypi_0
prometheus_client=0.14.1=py310hca03da5_0
prompt-toolkit=3.0.43=py310hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=3.20.3=py310h313beb8_0
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=14.0.2=py310h68338f1_0
pyautogen=0.2.27=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pydeck=0.8.0=py310hca03da5_2
pygments=2.15.1=py310hca03da5_1
pylint=2.16.2=py310hca03da5_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.14=hb885b13_1
python-dateutil=2.9.0post0=py310hca03da5_2
python-dotenv=1.0.1=pypi_0
python-fastjsonschema=2.16.2=py310hca03da5_0
python-json-logger=2.0.7=py310hca03da5_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hca03da5_0
pyyaml=6.0.1=py310h80987f9_0
pyzmq=25.1.2=py310h313beb8_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py310hca03da5_0
regex=2024.5.15=pypi_0
requests=2.32.2=py310hca03da5_0
rfc3339-validator=0.1.4=py310hca03da5_0
rfc3986-validator=0.1.1=py310hca03da5_0
rich=13.3.5=py310hca03da5_0
rpds-py=0.10.6=py310hf0e4da2_0
send2trash=1.8.2=py310hca03da5_0
setuptools=69.5.1=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.10=h313beb8_1
sniffio=1.3.1=pypi_0
soupsieve=2.5=py310hca03da5_0
sqlite=3.45.3=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.32.0=py310hca03da5_0
tenacity=8.2.2=py310hca03da5_0
termcolor=2.4.0=pypi_0
terminado=0.17.1=py310hca03da5_0
tiktoken=0.7.0=pypi_0
tinycss2=1.2.1=py310hca03da5_0
tk=8.6.14=h6ba3021_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py310hca03da5_0
tomlkit=0.11.1=py310hca03da5_0
toolz=0.12.0=py310hca03da5_0
tornado=6.3.3=py310h80987f9_0
tqdm=4.66.4=pypi_0
traitlets=5.7.1=py310hca03da5_0
typing-extensions=4.12.0=pypi_0
typing_extensions=4.11.0=py310hca03da5_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=py310hca03da5_0
utf8proc=2.6.1=h80987f9_1
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=py310hca03da5_0
wheel=0.43.0=py310hca03da5_0
widgetsnbextension=3.6.6=py310hca03da5_0
wrapt=1.14.1=py310h1a28f6b_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yfinance=0.2.40=pypi_0
zeromq=4.3.5=h313beb8_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
